extract `TradeResult` and `TradeSuccess`
